+ [fix] Modification and improvement of elevator csv
