+++ app/src/main/res/layout/layout_transfer_line.xml
@@ -34,6 +34,7 @@
                         android:id="@+id/transfer_line_circle_title"
                         android:layout_width="match_parent"
                         android:layout_height="match_parent"
+                        android:textSize="12sp"
                         android:gravity="center"
                         android:text="@string/test"
                         android:textColor="#FFFFFF"
